+++ changelog.md
@@ -4,7 +4,7 @@ Todos los cambios notables en este proyecto se documentarán en este archivo.
 El formato está basado en [Keep a Changelog](https://keepachangelog.com/es-ES/1.1.0/),
 y este proyecto sigue [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [Unreleased] - 2025-09-06
 ### Added
 - Caché alias de nodos conocidos.
 - Test de cache de alias conocidos.
@@ -13,6 +13,13 @@ y este proyecto sigue [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 ### Changed
 - Implementacion de propiedad ``is_connected``
 
-### Fixed
+## [Unreleaser] - 2025-09-13
+### Added
+- ``main.py``: Arranca servidor.
+- Clase ``OpcServer`` en ``opcua_server.py``.
+- Fuente nodos CSV.
 
+### Changued
+
+### Fixed